Add files via upload
In the ever-evolving landscape of the insurance industry, fraud detection has become a critical component for maintaining integrity and profitability. This Power BI project focuses on the development of a comprehensive dashboard that identifies and analyzes potential fraudulent claims in vehicle insurance. Leveraging advanced data analytics, visualization techniques and insights, this project aims to provide insurance companies with actionable intelligence to enhance their claims review process.